api-reference

+++ src/icpp/commands_build_native.py
@@ -3,18 +3,28 @@ import sys
 import subprocess
 import shutil
 import typer
+from typing_extensions import Annotated
 
 from icpp.__main__ import app
 from icpp import config_default
 from icpp.run_shell_cmd import run_shell_cmd
 
 from icpp.decorators import requires_native_compiler, requires_pro
+from icpp.options_build import to_compile_callback, option_to_compile_values_string
 
 
 @app.command()
 @requires_native_compiler()
-@requires_pro()
-def build_native() -> None:
+@requires_pro("build-native")
+def build_native(
+    to_compile: Annotated[
+        str,
+        typer.Option(
+            help=f"Files to compile {option_to_compile_values_string}.",
+            callback=to_compile_callback,
+        ),
+    ] = "all",
+) -> None:
     """Builds a native debug executable with your systems' Clang compiler.
 
     Reads 'icpp.toml' file in the current folder, and uses: \n
@@ -27,19 +37,20 @@ def build_native() -> None:
     build_path = icpp_toml.icpp_toml_path.parent / "build-native"
     typer.echo(f"Build folder: {build_path.resolve()}")
 
-    if build_path.exists():
-        typer.echo("Deleting the build-native folder.")
-        try:
-            shutil.rmtree(build_path)
-        except OSError as e:
-            typer.echo(f"ERROR: During delete of {build_path}:\n {e.strerror}")
+    if to_compile == "all":
+        if build_path.exists():
+            typer.echo("Deleting the build-native folder.")
+            try:
+                shutil.rmtree(build_path)
+            except OSError as e:
+                typer.echo(f"ERROR: During delete of {build_path}:\n {e.strerror}")
 
     if not build_path.exists():
         build_path.mkdir()
 
     try:
         # ----------------------------------------------------------------------
-        # compile all C++ files, if we have any
+        # compile 'mine' C++ files, if we have any
         cpp_files = (
             icpp_toml.build_native["cpp_files"] + icpp_toml.build_wasm["cpp_files"]
         )
@@ -57,7 +68,7 @@ def build_native() -> None:
             run_shell_cmd(cmd, cwd=build_path)
 
         # ----------------------------------------------------------------------
-        # compile all C files, if we have any
+        # compile 'mine' C files, if we have any
         c_files = icpp_toml.build_native["c_files"] + icpp_toml.build_wasm["c_files"]
         c_compile_flags_s = icpp_toml.build_native["c_compile_flags_s"]
         if len(c_files.strip()) > 0:
@@ -72,33 +83,34 @@ def build_native() -> None:
             typer.echo(cmd)
             run_shell_cmd(cmd, cwd=build_path)
 
-        # ----------------------------------------------------------------------
-        # compile the C++ files of the Mock IC, if we have any
-        if len(config_default.MOCKIC_CPP_FILES.strip()) > 0:
-            cmd = (
-                f"{config_default.NATIVE_CPP} "
-                f"{config_default.NATIVE_CPPFLAGS} "
-                f"-c {config_default.MOCKIC_CPP_FILES}"
-            )
+        if to_compile == "all":
+            # ----------------------------------------------------------------------
+            # compile the C++ files of the Mock IC, if we have any
+            if len(config_default.MOCKIC_CPP_FILES.strip()) > 0:
+                cmd = (
+                    f"{config_default.NATIVE_CPP} "
+                    f"{config_default.NATIVE_CPPFLAGS} "
+                    f"-c {config_default.MOCKIC_CPP_FILES}"
+                )
 
-            typer.echo("--")
-            typer.echo("Compiling C++ files of the Mock IC with command:")
-            typer.echo(cmd)
-            run_shell_cmd(cmd, cwd=build_path)
+                typer.echo("--")
+                typer.echo("Compiling C++ files of the Mock IC with command:")
+                typer.echo(cmd)
+                run_shell_cmd(cmd, cwd=build_path)
 
-        # ----------------------------------------------------------------------
-        # compile the C files of the Mock IC, if we have any
-        if len(config_default.MOCKIC_C_FILES.strip()) > 0:
-            cmd = (
-                f"{config_default.NATIVE_C} "
-                f"{config_default.NATIVE_CFLAGS} "
-                f"-c {config_default.MOCKIC_C_FILES}"
-            )
+            # ----------------------------------------------------------------------
+            # compile the C files of the Mock IC, if we have any
+            if len(config_default.MOCKIC_C_FILES.strip()) > 0:
+                cmd = (
+                    f"{config_default.NATIVE_C} "
+                    f"{config_default.NATIVE_CFLAGS} "
+                    f"-c {config_default.MOCKIC_C_FILES}"
+                )
 
-            typer.echo("--")
-            typer.echo("Compiling C files of the Mock IC with command:")
-            typer.echo(cmd)
-            run_shell_cmd(cmd, cwd=build_path)
+                typer.echo("--")
+                typer.echo("Compiling C files of the Mock IC with command:")
+                typer.echo(cmd)
+                run_shell_cmd(cmd, cwd=build_path)
 
         # ----------------------------------------------------------------------
         # link it into a native executable
